Make truncate work with unicode
Add test cases for some edge cases and japanese characters
